Move Build episode list out from string builder
